+++ src/models/User/User.t.ts
@@ -1,0 +1,6 @@
+export interface IUser {
+	id?: string; // uuid
+	username: string;
+	age: number;
+	hobbies: string[];
+}
